requirements.txt
tensorflow==2.13.0
keras==2.13.1
fastapi==0.103.2
uvicorn==0.33.0
pillow
numpy
h5py
streamlit==1.40.1